+++ src/types/index.ts
@@ -7,3 +7,4 @@ export * from './Fetcher';
 export * from './RestStdOptions';
 export * from './Encryption';
 export * from './Csrf';
+export * from './modelKeys';